Fix complexity
