+++ src/Service/Matcher/DateBeforeTermMatcher.php
@@ -1,0 +1,24 @@
+<?php
+declare(strict_types=1);
+
+namespace FD\LogViewer\Service\Matcher;
+
+use FD\LogViewer\Entity\Expression\DateBeforeTerm;
+use FD\LogViewer\Entity\Expression\TermInterface;
+use FD\LogViewer\Entity\Index\LogRecord;
+
+/**
+ * @implements TermMatcherInterface<DateBeforeTerm>
+ */
+class DateBeforeTermMatcher implements TermMatcherInterface
+{
+    public function supports(TermInterface $term): bool
+    {
+        return $term instanceof DateBeforeTerm;
+    }
+
+    public function matches(TermInterface $term, LogRecord $record): bool
+    {
+        return $record->date <= $term->date->getTimestamp();
+    }
+}
