+++ tests/Integration/Service/File/LogParserTest.php
@@ -7,8 +7,8 @@ use FD\LogViewer\Entity\Output\DirectionEnum;
 use FD\LogViewer\Entity\Request\LogQueryDto;
 use FD\LogViewer\Reader\Stream\StreamReaderFactory;
 use FD\LogViewer\Service\File\LogParser;
-use FD\LogViewer\Service\File\LogRecordMatcher;
 use FD\LogViewer\Service\File\Monolog\MonologLineParser;
+use FD\LogViewer\Service\Matcher\LogRecordMatcher;
 use FD\LogViewer\Tests\Integration\AbstractIntegrationTestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
 use Psr\Clock\ClockInterface;
