+++ tests/Unit/Iterator/LogRecordFilterIteratorTest.php
@@ -7,7 +7,7 @@ use ArrayIterator;
 use FD\LogViewer\Entity\Expression\Expression;
 use FD\LogViewer\Entity\Index\LogRecord;
 use FD\LogViewer\Iterator\LogRecordFilterIterator;
-use FD\LogViewer\Service\File\LogRecordMatcher;
+use FD\LogViewer\Service\Matcher\LogRecordMatcher;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
